Dev (#2)
* Use public freeipa-issuer-controller-manager repo

Signed-off-by: Bogdan Dobrelia <bogdan.dobrelia@spearline.com>

* Fix code-owners

Signed-off-by: Bogdan Dobrelia <bogdan.dobrelia@spearline.com>

* Allow templates merged with extra_values

For 'extra' values provider, allow it taking templates with
dynamically evaluated variables as ECR containers images data
provides it. Then merge it with extra_values by keys as it is
described in extra_tpl and/or helm-dependencies:

This example results in the container spec args containing:
  - --acme-http01-solver-image=<rewritten_ECR_image_repo>:<image_tag>

helm-dependencies.yaml:
  args.acme-http01-solver-image:
    manager: extra
    name:
      # gets rewritten as <rewritten_ECR_image_repo> by images prep
      repository: <image_repo>
    ver:
      tag: <image_tag>
  extraArgs.someimagevar:
    manager: extra
    name:
      repository: ...

terragrunt.yml:
cert-manager = {
  extra_tpl = {
    args = [
      # note the $foo format, it will be rendered as ${foo} tpl var.
      "--acme-http01-solver-image=$acme-http01-solver-image-repo:$acme-http01-solver-image-repo-tag"
    ],
    extraArgs = [... --foo=$someimagevar-repo ... ]
  }
}
whereis rendered template values for ${acme-http01-solver-image-repo} and
${acme-http01-solver-image-repo-tag} vars become autodiscovered by the
extra provider, based on provided data in helm-dependencies. Also note
that 'args' key name must match each other in both places.

NOTE: both extra_tpl and extra_values cannot overlap, so 'args' and
'extraArgs' must be defined in extra_values (if no templates in it),
or in extra_tpl, if containst templates, but not in both places.

Signed-off-by: Bogdan Dobrelia <bogdan.dobrelia@spearline.com>

* Drop PR title format CI checks

Signed-off-by: Bogdan Dobrelia <bogdan.dobrelia@spearline.com>

* Make cert-manager backend configurable

Unhardcode letsencrypt, provide options for provider, server

Signed-off-by: Bogdan Dobrelia <bogdan.dobrelia@spearline.com>

Signed-off-by: Bogdan Dobrelia <bogdan.dobrelia@spearline.com>