Make storage classes reclaim policy configurable
For EBS and EFS storage classes that we configure for tenants
ensure unbound PVs may be retained, after deletion of resources that
use it. Keep the default value unchanged (which is to
Delete those).

Create an additional VolumeSnapshotClass for EFS CSI, and mark it
to become the default class vor snapshots.

On tEKS side we need to decide with which reclaim policy to deploy
those. Retain is safer for tenants, although that requires
cleaning up policies developed and explained to tenants.

Signed-off-by: Bogdan Dobrelia <bogdan.dobrelia@spearline.com>